+++ sphinxcontrib/mat_types.py
@@ -214,8 +214,10 @@ def analyze(app):
         )
         return
 
-    # Interpret `matlab_src_dir` relative to the `conf.py` file.
-    basedir = os.path.normpath(os.path.join(app.confdir, app.env.config.matlab_src_dir))
+    # Interpret `matlab_src_dir` relative to the sphinx source directory.
+    basedir = os.path.normpath(
+        os.path.join(app.env.srcdir, app.env.config.matlab_src_dir)
+    )
     MatObject.basedir = basedir  # set MatObject base directory
     MatObject.sphinx_env = app.env  # pass env to MatObject cls
     MatObject.sphinx_app = app  # pass app to MatObject cls